add loaders, add alerts to crud actions, add alerts handling errors

+++ klient/src/Components/CarManager/CarManager.jsx
@@ -3,37 +3,71 @@ import { Button } from "reactstrap";
 import { Link } from "react-router-dom";
 import CarManagerTable from "../CarManagerTable/CarManagerTable";
 import axios from "axios";
+import Loader from "react-loader-spinner";
+import { useHistory } from "react-router-dom";
+import Swal from "sweetalert2";
 
 export default function CarManager() {
   const [data, setData] = useState([]);
-
+  const [isLoading, setIsLoading] = useState(false);
+  let history = useHistory();
   const fetchCars = async () => {
     try {
+      setIsLoading(true);
       const response = await axios.get("https://localhost:44390/api/cars");
       setData(response.data);
-    } catch (e) {
-      console.log(e);
-      setData(data);
+      setIsLoading(false);
+    } catch (error) {
+      Swal.fire("Oops...", "Something went wrong!", "error").then(() =>
+        history.goBack()
+      );
     }
   };
   useEffect(() => {
     fetchCars();
   }, []);
 
-  async function deleteCar(id) {
-    await axios({
-      url: "https://localhost:44390/api/cars/" + id,
-      method: "DELETE",
+  const swalWithBootstrapButtons = Swal.mixin({
+    customClass: {
+      confirmButton: 'btn btn-success',
+      cancelButton: 'btn btn-danger'
+    },
+    buttonsStyling: false
+  })
+
+  function deleteCar(id) {
+    swalWithBootstrapButtons.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, cancel!',
+      reverseButtons: true
+    }).then((result) => {
+      if (result.value) {
+        axios({
+          url: "https://localhost:44390/api/cars/" + id,
+          method: "DELETE",
+        }).catch((error) => {
+            Swal.fire("Oops", "Something went wrong when deleting. Error:"+error.response, "error");
+          });
+        swalWithBootstrapButtons.fire(
+          'Deleted!',
+          'Your car has been deleted.',
+          'success'
+        )
+        setTimeout(()=>fetchCars(),2000);
+      } else if (
+        result.dismiss === Swal.DismissReason.cancel
+      ) {
+        swalWithBootstrapButtons.fire(
+          'Cancelled',
+          'Your car is safe :)',
+          'error'
+        )
+      }
     })
-      .then((res) => {
-        if (res.status === 200) {
-          alert("Deleted");
-          window.location.reload(false);
-        }
-      })
-      .catch((error) => {
-        console.log(error.response);
-      });
   }
 
   const columns = useMemo(
@@ -85,5 +119,15 @@ export default function CarManager() {
     []
   );
 
-  return <CarManagerTable columns={columns} data={data} />;
+  return (
+    <div>
+      {isLoading ? (
+        <div className="loader">
+          <Loader type="Oval" color="#00BFFF" height={80} width={80} />
+        </div>
+      ) : (
+        <CarManagerTable columns={columns} data={data} />
+      )}
+    </div>
+  );
 }
